requirements.txt
python-dotenv
PyMuPDF
gigachat
gigachain
sentence-transformers
faiss-cpu
